cache user full names and nitroduced group based flushing
git-svn-id: https://svn.liip.ch/repos/public/fluxcms/trunk@6774 99f7f646-bdd5-0310-89f9-ddcb10b10bf3

+++ inc/bx/cache/memcache.php
@@ -1,11 +1,11 @@
 <?php
 
 class bx_cache_memcache {
-	protected $cache;
-	static protected $instance = NULL;
+    protected $cache;
+    static protected $instance = NULL;
     
-	private function __construct() {
-		$this->cache = new Memcache();
+    private function __construct() {
+        $this->cache = new Memcache();
         
         $servers = $GLOBALS['POOL']->config->cacheOptions;
         if ($servers) {
@@ -18,52 +18,70 @@ class bx_cache_memcache {
             }
         }
         //bx_helpers_debug::webdump($this->cache->getExtendedStats());
-	}
+    }
 
-	public static function getInstance() {
-		if (!isset(self::$instance)) {
-			self::$instance = new bx_cache_memcache;
-		}
+    public static function getInstance() {
+        if (!isset(self::$instance)) {
+            self::$instance = new bx_cache_memcache;
+        }
 
-		return self::$instance;
-	}
+        return self::$instance;
+    }
 
-	public function connect( $server, $port = 11211, $per = true, $weight = 1, $timeout = 1, $retry = 15 ) {
+    public function connect( $server, $port = 11211, $per = true, $weight = 1, $timeout = 1, $retry = 15 ) {
        
-		return $this->cache->addServer($server, $port, $per, $weight, $timeout, $retry);
-	}
+        return $this->cache->addServer($server, $port, $per, $weight, $timeout, $retry);
+    }
 
-	public function add($key, $val, $expires = null, $group = null) {
-        //error_log("add " . $key);
-		return $this->cache->add($key, $val, false, $expires);
-	}
-
-	public function del($key) {
-		return $this->cache->delete($key);
-	}
-
-	public function flush() {
-		return $this->cache->flush();
-	}
-
-	public function get($keys) {
-        //error_log("get " . $keys);
-		return $this->cache->get($keys);
-	}
-
-	public function replace($key, $val, $expires = 25200, $group = null) {
-		return $this->cache->replace($key, $val, false, $expire);
-	}
-
-	public function set($key, $val,  $expires = 25200, $group = null) {
-        
+    public function set($key, $val,  $expires = 25200, $group = null) {
+        if ($group) {
+            $keys = $this->cache->get($group);
+            if (!in_array($key,$keys)) {
+                $keys[] = $key;
+                $this->cache->set($group,$keys);
+            }
+        }
         return $this->cache->set($key, $val, false, $expires);
-	}
+    }
      
+    
+ /*   public function add($key, $val, $expires = null, $group = null) {
+        //error_log("add " . $key);
+        return $this->cache->add($key, $val, false, $expires);
+    }
+      public function replace($key, $val, $expires = 25200, $group = null) {
+        return $this->cache->replace($key, $val, false, $expire);
+    }
+
+    
+    */
+
+    public function del($key) {
+        return $this->cache->delete($key);
+    }
+
+    public function flush($group = null) {
+        if ($group) {
+            $keys = $this->cache->get($group);
+            foreach ($keys as $key) {
+                $this->cache->delete($key);
+            }
+        } else {
+            return $this->cache->flush();
+        }
+    }
+
+    public function get($keys) {
+        //error_log("get " . $keys);
+        return $this->cache->get($keys);
+    }
+
+  
+  
    public function getStats() {
 
-		return $this->cache->getExtendedStats();
-	}
+        return $this->cache->getExtendedStats();
+    }
 
 }
